requirements.txt
# Core dependencies
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<0.31
pydantic>=2.0,<3

# Database
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2

# Document processing
docling>=2.0,<3
langchain-text-splitters>=0.1,<1

# External APIs
requests>=2.31,<3
google-generativeai>=0.8.0

# Utilities
python-multipart>=0.0.6

# Image processing
Pillow>=10.0,<11
